working atlas v1(no search implemented)

+++ modules/atlas-addon/shared/Messages.ts
@@ -15,7 +15,7 @@ export class itemLootPacket {
     @MsgPrimitive
         itemCountMax: uint32 = 0;
     @MsgPrimitive
-        dropChance: uint32 = 0;  
+        dropChance: double = 0;  
 }
 
 @Message
